# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API Documentation",
    "version": "1.0.0",
    "description": "API documentation for my Next.js app"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://book-teller-batrinutsilvius-projects.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token from the login endpoint"
      }
    }
  },
  "paths": {
...
    "/api/books/{book_id}/languages": {
      "get": {
        "summary": "Gets all translations of a book",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The book ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of book translations with language info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "BookTranslation ID"
                      },
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string",
                        "nullable": true
                      },
                      "language": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "country_code": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/books/categories/{category_id}/languages/{language_id}": {
      "get": {
        "summary": "Gets all the books from a category by language",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The category ID"
          },
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The language ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of books in the category with translation for the given language",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "BookCategory ID"
                      },
                      "book": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "photo_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "bookTranslations": {
                            "type": "array",
                            "description": "Contains one translation for the requested language, sorted by title",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "integer"
                                },
                                "title": {
                                  "type": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}